Put skill panel specific code into panel class. Remove from collapsing panel so it doesnt affect glossary use
